fix acne on cylinder caps

+++ src/object/cylinder.rs
@@ -20,9 +20,9 @@ impl Shape for Cylinder {
     fn object_normal_at(&self, point: Point3D, _uv: Option<(f64, f64)>) -> Normal3D {
         let distance = point.x().powi(2) + point.z().powi(2);
 
-        if distance < 1.0 && point.y() >= self.max_y - f64::EPSILON {
+        if distance < 1.0 && point.y() >= self.max_y - (f32::EPSILON as f64) {
             Normal3D::POSITIVE_Y
-        } else if distance < 1.0 && point.y() <= self.min_y + f64::EPSILON {
+        } else if distance < 1.0 && point.y() <= self.min_y + (f32::EPSILON as f64) {
             Normal3D::NEGATIVE_Y
         } else {
             Vector3D::new(point.x(), 0.0, point.z()).normalised()
